backend/requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database - latest versions for Python 3.14 support
sqlalchemy>=2.0.36
alembic>=1.14.0
psycopg[binary]>=3.2.0

# Cache & Queue
redis>=5.0.1
celery>=5.3.6

# Data Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Environment
python-dotenv>=1.0.0

# HTTP Client
httpx>=0.26.0

# Security
cryptography>=41.0.0

# Utilities
python-dateutil>=2.8.2
